+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/InlineQueries/InlineQueryResult/abstracts/results/mpeg4gif/InlineQueryResultMpeg4GifCommon.kt
@@ -1,11 +1,11 @@
 package com.github.insanusmokrassar.TelegramBotAPI.types.InlineQueries.InlineQueryResult.abstracts.results.mpeg4gif
 
-import com.github.insanusmokrassar.TelegramBotAPI.CommonAbstracts.Captioned
+import com.github.insanusmokrassar.TelegramBotAPI.CommonAbstracts.CaptionedOutput
 import com.github.insanusmokrassar.TelegramBotAPI.types.InlineQueries.InlineQueryResult.abstracts.*
 
 const val inlineQueryResultMpeg4GifType = "mpeg4_gif"
 
 interface InlineQueryResultMpeg4GifCommon : InlineQueryResult,
     OptionallyTitledInlineQueryResult,
-    Captioned,
+    CaptionedOutput,
     WithInputMessageContentInlineQueryResult