+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/InlineQueries/InlineQueryResult/abstracts/results/photo/InlineQueryResultPhotoCommon.kt
@@ -1,6 +1,6 @@
 package com.github.insanusmokrassar.TelegramBotAPI.types.InlineQueries.InlineQueryResult.abstracts.results.photo
 
-import com.github.insanusmokrassar.TelegramBotAPI.CommonAbstracts.Captioned
+import com.github.insanusmokrassar.TelegramBotAPI.CommonAbstracts.CaptionedOutput
 import com.github.insanusmokrassar.TelegramBotAPI.types.InlineQueries.InlineQueryResult.abstracts.*
 
 const val inlineQueryResultPhotoType = "photo"
@@ -8,5 +8,5 @@ const val inlineQueryResultPhotoType = "photo"
 interface InlineQueryResultPhotoCommon : InlineQueryResult,
     OptionallyTitledInlineQueryResult,
     DescribedInlineQueryResult,
-    Captioned,
+    CaptionedOutput,
     WithInputMessageContentInlineQueryResult