+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/message/content/media/VideoContent.kt
@@ -19,7 +19,7 @@ data class VideoContent(
     override val media: VideoFile,
     override val caption: String? = null,
     override val captionEntities: List<MessageEntity> = emptyList()
-) : CaptionedMediaContent, MediaGroupContent {
+) : MediaGroupContent {
     override fun createResend(
         chatId: ChatIdentifier,
         disableNotification: Boolean,